+++ core/src/commonMain/kotlin/com/powersync/db/PowerSyncDatabaseImpl.kt
@@ -14,6 +14,7 @@ import com.powersync.db.crud.CrudRow
 import com.powersync.db.crud.CrudTransaction
 import com.powersync.db.internal.InternalDatabaseImpl
 import com.powersync.db.internal.InternalTable
+import com.powersync.db.internal.PowerSyncVersion
 import com.powersync.db.schema.Schema
 import com.powersync.db.schema.toSerializable
 import com.powersync.sync.PriorityStatusEntry
@@ -217,21 +218,7 @@ internal class PowerSyncDatabaseImpl(
             }
 
             launch {
-                stream.status.asFlow().collect {
-                    currentStatus.update(
-                        connected = it.connected,
-                        connecting = it.connecting,
-                        uploading = it.uploading,
-                        downloading = it.downloading,
-                        lastSyncedAt = it.lastSyncedAt,
-                        hasSynced = it.hasSynced,
-                        uploadError = it.uploadError,
-                        downloadError = it.downloadError,
-                        clearDownloadError = it.downloadError == null,
-                        clearUploadError = it.uploadError == null,
-                        priorityStatusEntries = it.priorityStatusEntries,
-                    )
-                }
+                currentStatus.trackOther(stream.status)
             }
 
             launch {
@@ -258,10 +245,10 @@ internal class PowerSyncDatabaseImpl(
             return null
         }
 
-        val entries =
+        var entries =
             internalDb.getAll(
                 "SELECT id, tx_id, data FROM ps_crud ORDER BY id ASC LIMIT ?",
-                listOf(limit.toLong()),
+                listOf(limit.toLong() + 1),
             ) {
                 CrudEntry.fromRow(
                     CrudRow(
@@ -278,7 +265,7 @@ internal class PowerSyncDatabaseImpl(
 
         val hasMore = entries.size > limit
         if (hasMore) {
-            entries.dropLast(entries.size - limit)
+            entries = entries.dropLast(1)
         }
 
         return CrudBatch(entries, hasMore, complete = { writeCheckpoint ->
@@ -351,11 +338,12 @@ internal class PowerSyncDatabaseImpl(
     override fun onChange(
         tables: Set<String>,
         throttleMs: Long,
+        triggerImmediately: Boolean,
     ): Flow<Set<String>> =
         flow {
             waitReady()
             emitAll(
-                internalDb.onChange(tables, throttleMs),
+                internalDb.onChange(tables, throttleMs, triggerImmediately),
             )
         }
 
@@ -436,11 +424,14 @@ internal class PowerSyncDatabaseImpl(
             syncSupervisorJob = null
         }
 
-        currentStatus.update(
-            connected = false,
-            connecting = false,
-            lastSyncedAt = currentStatus.lastSyncedAt,
-        )
+        currentStatus.update {
+            copy(
+                connected = false,
+                connecting = false,
+                downloading = false,
+                downloadProgress = null,
+            )
+        }
     }
 
     override suspend fun disconnectAndClear(clearLocal: Boolean) {
@@ -449,7 +440,7 @@ internal class PowerSyncDatabaseImpl(
         internalDb.writeTransaction { tx ->
             tx.getOptional("SELECT powersync_clear(?)", listOf(if (clearLocal) "1" else "0")) {}
         }
-        currentStatus.update(lastSyncedAt = null, hasSynced = false)
+        currentStatus.update { copy(lastSyncedAt = null, hasSynced = false) }
     }
 
     private suspend fun updateHasSynced() {
@@ -489,11 +480,13 @@ internal class PowerSyncDatabaseImpl(
             }
         }
 
-        currentStatus.update(
-            hasSynced = lastSyncedAt != null,
-            lastSyncedAt = lastSyncedAt,
-            priorityStatusEntries = priorityStatus,
-        )
+        currentStatus.update {
+            copy(
+                hasSynced = lastSyncedAt != null,
+                lastSyncedAt = lastSyncedAt,
+                priorityStatusEntries = priorityStatus,
+            )
+        }
     }
 
     override suspend fun waitForFirstSync() = waitForFirstSyncImpl(null)
@@ -533,20 +526,9 @@ internal class PowerSyncDatabaseImpl(
      * Check that a supported version of the powersync extension is loaded.
      */
     private fun checkVersion(powerSyncVersion: String) {
-        // Parse version
-        val versionInts: List<Int> =
-            try {
-                powerSyncVersion
-                    .split(Regex("[./]"))
-                    .take(3)
-                    .map { it.toInt() }
-            } catch (e: Exception) {
-                throw Exception("Unsupported powersync extension version. Need ^0.2.0, got: $powerSyncVersion. Details: $e")
-            }
-
-        // Validate ^0.2.0
-        if (versionInts[0] != 0 || versionInts[1] < 2 || versionInts[2] < 0) {
-            throw Exception("Unsupported powersync extension version. Need ^0.2.0, got: $powerSyncVersion")
+        val version = PowerSyncVersion.parse(powerSyncVersion)
+        if (version < PowerSyncVersion.MINIMUM) {
+            PowerSyncVersion.mismatchError(powerSyncVersion)
         }
     }
 }