+++ core/src/commonMain/kotlin/com/powersync/sync/SyncStream.kt
@@ -72,7 +72,7 @@ internal class SyncStream(
         var invalidCredentials = false
         clientId = bucketStorage.getClientId()
         while (true) {
-            status.update(connecting = true)
+            status.update { copy(connecting = true) }
             try {
                 if (invalidCredentials) {
                     // This may error. In that case it will be retried again on the next
@@ -87,15 +87,9 @@ internal class SyncStream(
                 }
 
                 logger.e("Error in streamingSync: ${e.message}")
-                status.update(
-                    downloadError = e,
-                )
+                status.update { copy(downloadError = e) }
             } finally {
-                status.update(
-                    connected = false,
-                    connecting = true,
-                    downloading = false,
-                )
+                status.update { copy(connected = false, connecting = true, downloading = false) }
                 delay(retryDelayMs)
             }
         }
@@ -142,7 +136,7 @@ internal class SyncStream(
                     }
 
                     checkedCrudItem = nextCrudItem
-                    status.update(uploading = true)
+                    status.update { copy(uploading = true) }
                     uploadCrud()
                 } else {
                     // Uploading is completed
@@ -150,7 +144,7 @@ internal class SyncStream(
                     break
                 }
             } catch (e: Exception) {
-                status.update(uploading = false, uploadError = e)
+                status.update { copy(uploading = false, uploadError = e) }
 
                 if (e is CancellationException) {
                     throw e
@@ -161,7 +155,7 @@ internal class SyncStream(
                 break
             }
         }
-        status.update(uploading = false)
+        status.update { copy(uploading = false) }
     }
 
     private suspend fun getWriteCheckpoint(): String {
@@ -215,7 +209,7 @@ internal class SyncStream(
                     throw RuntimeException("Received error when connecting to sync stream: ${httpResponse.bodyAsText()}")
                 }
 
-                status.update(connected = true, connecting = false)
+                status.update { copy(connected = true, connecting = false) }
                 val channel: ByteReadChannel = httpResponse.body()
 
                 while (!channel.isClosedForRead) {
@@ -260,7 +254,7 @@ internal class SyncStream(
             }
         }
 
-        status.update(downloading = false)
+        status.update { abortedDownload() }
 
         return state
     }
@@ -294,7 +288,6 @@ internal class SyncStream(
     ): SyncStreamState {
         val (checkpoint) = line
         state.targetCheckpoint = checkpoint
-        status.update(downloading = true)
 
         val bucketsToDelete = state.bucketSet!!.toMutableList()
         val newBuckets = mutableSetOf<String>()
@@ -306,15 +299,30 @@ internal class SyncStream(
             }
         }
 
-        if (bucketsToDelete.size > 0) {
+        state.bucketSet = newBuckets
+        startTrackingCheckpoint(checkpoint, bucketsToDelete)
+
+        return state
+    }
+
+    private suspend fun startTrackingCheckpoint(
+        checkpoint: Checkpoint,
+        bucketsToDelete: List<String>,
+    ) {
+        val progress = bucketStorage.getBucketOperationProgress()
+        status.update {
+            copy(
+                downloading = true,
+                downloadProgress = SyncDownloadProgress(progress, checkpoint),
+            )
+        }
+
+        if (bucketsToDelete.isNotEmpty()) {
             logger.i { "Removing buckets [${bucketsToDelete.joinToString(separator = ", ")}]" }
         }
 
-        state.bucketSet = newBuckets
         bucketStorage.removeBuckets(bucketsToDelete)
         bucketStorage.setTargetCheckpoint(checkpoint)
-
-        return state
     }
 
     private suspend fun handleStreamingSyncCheckpointComplete(state: SyncStreamState): SyncStreamState {
@@ -343,12 +351,7 @@ internal class SyncStream(
             logger.i { "validated checkpoint ${state.appliedCheckpoint}" }
 
             state.validatedCheckpoint = state.targetCheckpoint
-            status.update(
-                lastSyncedAt = Clock.System.now(),
-                downloading = false,
-                hasSynced = true,
-                clearDownloadError = true,
-            )
+            status.update { copyWithCompletedDownload() }
         } else {
             logger.d { "Could not apply checkpoint. Waiting for next sync complete line" }
         }
@@ -376,20 +379,22 @@ internal class SyncStream(
             logger.i { "validated partial checkpoint ${state.appliedCheckpoint} up to priority of $priority" }
         }
 
-        status.update(
-            priorityStatusEntries =
-                buildList {
-                    // All states with a higher priority can be deleted since this partial sync includes them.
-                    addAll(status.priorityStatusEntries.filter { it.priority >= line.priority })
-                    add(
-                        PriorityStatusEntry(
-                            priority = priority,
-                            lastSyncedAt = Clock.System.now(),
-                            hasSynced = true,
-                        ),
-                    )
-                },
-        )
+        status.update {
+            copy(
+                priorityStatusEntries =
+                    buildList {
+                        // All states with a higher priority can be deleted since this partial sync includes them.
+                        addAll(status.priorityStatusEntries.filter { it.priority >= line.priority })
+                        add(
+                            PriorityStatusEntry(
+                                priority = priority,
+                                lastSyncedAt = Clock.System.now(),
+                                hasSynced = true,
+                            ),
+                        )
+                    },
+            )
+        }
         return state
     }
 
@@ -401,8 +406,6 @@ internal class SyncStream(
         if (state.targetCheckpoint == null) {
             throw Exception("Checkpoint diff without previous checkpoint")
         }
-
-        status.update(downloading = true)
 
         val newBuckets = mutableMapOf<String, BucketChecksum>()
 
@@ -423,15 +426,7 @@ internal class SyncStream(
             )
 
         state.targetCheckpoint = newCheckpoint
-
-        state.bucketSet = newBuckets.keys.toMutableSet()
-
-        val bucketsToDelete = checkpointDiff.removedBuckets
-        if (bucketsToDelete.isNotEmpty()) {
-            logger.d { "Remove buckets $bucketsToDelete" }
-        }
-        bucketStorage.removeBuckets(bucketsToDelete)
-        bucketStorage.setTargetCheckpoint(state.targetCheckpoint!!)
+        startTrackingCheckpoint(newCheckpoint, checkpointDiff.removedBuckets)
 
         return state
     }
@@ -440,8 +435,9 @@ internal class SyncStream(
         data: SyncLine.SyncDataBucket,
         state: SyncStreamState,
     ): SyncStreamState {
-        status.update(downloading = true)
-        bucketStorage.saveSyncData(SyncDataBatch(listOf(data)))
+        val batch = SyncDataBatch(listOf(data))
+        status.update { copy(downloading = true, downloadProgress = downloadProgress?.incrementDownloaded(batch)) }
+        bucketStorage.saveSyncData(batch)
         return state
     }