+++ demos/supabase-todolist/shared/src/commonMain/kotlin/com/powersync/demos/App.kt
@@ -5,7 +5,6 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -13,7 +12,7 @@ import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Modifier
 import com.powersync.DatabaseDriverFactory
 import com.powersync.PowerSyncDatabase
-import com.powersync.bucket.BucketPriority
+import com.powersync.compose.composeState
 import com.powersync.connector.supabase.SupabaseConnector
 import com.powersync.connectors.PowerSyncBackendConnector
 import com.powersync.demos.components.EditDialog
@@ -25,7 +24,6 @@ import com.powersync.demos.screens.HomeScreen
 import com.powersync.demos.screens.SignInScreen
 import com.powersync.demos.screens.SignUpScreen
 import com.powersync.demos.screens.TodosScreen
-import kotlinx.coroutines.flow.debounce
 import kotlinx.coroutines.runBlocking
 import org.koin.compose.KoinApplication
 import org.koin.compose.koinInject
@@ -71,19 +69,8 @@ fun AppContent(
     db: PowerSyncDatabase = koinInject(),
     modifier: Modifier = Modifier,
 ) {
-    // Debouncing the status flow prevents flicker
-    val status by db.currentStatus
-        .asFlow()
-        .debounce(200)
-        .collectAsState(initial = db.currentStatus)
 
-    // This assumes that the buckets for lists has a priority of 1 (but it will work fine with sync
-    // rules not defining any priorities at all too). When giving lists a higher priority than
-    // items, we can have a consistent snapshot of lists without items. In the case where many items
-    // exist (that might take longer to sync initially), this allows us to display lists earlier.
-    val hasSyncedLists by remember {
-        derivedStateOf { status.statusForPriority(BucketPriority(1)).hasSynced }
-    }
+    val status by db.currentStatus.composeState()
 
     val authViewModel = koinViewModel<AuthViewModel>()
     val navController = koinInject<NavController>()